+++ lib/src/shape/arrow.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
 
-class ArrowClipper extends CustomClipper<Path> {
-  @override
-  Path getClip(Size size) {
+mixin ArrowShape {
+  Path getArrowPath(Size size) {
     final Path path = Path();
     path.moveTo(0, 0);
     path.lineTo(size.width * 3.5 / 4, 0.0);
@@ -12,7 +11,4 @@ class ArrowClipper extends CustomClipper<Path> {
     path.close();
     return path;
   }
-
-  @override
-  bool shouldReclip(ArrowClipper oldClipper) => this != oldClipper;
 }