+++ lib/src/shape/chevron_crumb.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
 
-class ChevronClipper extends CustomClipper<Path> {
-  @override
-  Path getClip(Size size) {
+mixin Chevron {
+  Path getChevronPath(Size size) {
     final Path path = Path();
     path.moveTo(0, 0);
     path.lineTo(size.width * 3.5 / 4, 0.0);
@@ -13,7 +12,4 @@ class ChevronClipper extends CustomClipper<Path> {
     path.close();
     return path;
   }
-
-  @override
-  bool shouldReclip(ChevronClipper oldClipper) => this != oldClipper;
 }